added devtools support for ssl cert testing

+++ docker/openemr/flex-3.12/autoconfig.sh
@@ -84,10 +84,13 @@ if [ -f /etc/docker-leader ] ||
         -days 365 -nodes \
         -subj "/C=xx/ST=x/L=x/O=x/OU=x/CN=localhost"
     fi
-    rm -f /etc/ssl/certs/webserver.cert.pem
-    rm -f /etc/ssl/private/webserver.key.pem
-    ln -s /etc/ssl/certs/selfsigned.cert.pem /etc/ssl/certs/webserver.cert.pem
-    ln -s /etc/ssl/private/selfsigned.key.pem /etc/ssl/private/webserver.key.pem
+    if [ ! -f /etc/ssl/docker-selfsigned-configured ]; then
+        rm -f /etc/ssl/certs/webserver.cert.pem
+        rm -f /etc/ssl/private/webserver.key.pem
+        ln -s /etc/ssl/certs/selfsigned.cert.pem /etc/ssl/certs/webserver.cert.pem
+        ln -s /etc/ssl/private/selfsigned.key.pem /etc/ssl/private/webserver.key.pem
+        touch /etc/ssl/docker-selfsigned-configured
+    fi
 
     if [ "$DOMAIN" != "" ]; then
         if [ "$EMAIL" != "" ]; then
@@ -106,10 +109,13 @@ if [ -f /etc/docker-leader ] ||
         fi
 
         # run letsencrypt as a daemon and reference the correct cert
-        rm -f /etc/ssl/certs/webserver.cert.pem
-        rm -f /etc/ssl/private/webserver.key.pem
-        ln -s /etc/letsencrypt/live/$DOMAIN/fullchain.pem /etc/ssl/certs/webserver.cert.pem
-        ln -s /etc/letsencrypt/live/$DOMAIN/privkey.pem /etc/ssl/private/webserver.key.pem
+        if [ ! -f /etc/ssl/docker-letsencrypt-configured ]; then
+            rm -f /etc/ssl/certs/webserver.cert.pem
+            rm -f /etc/ssl/private/webserver.key.pem
+            ln -s /etc/letsencrypt/live/$DOMAIN/fullchain.pem /etc/ssl/certs/webserver.cert.pem
+            ln -s /etc/letsencrypt/live/$DOMAIN/privkey.pem /etc/ssl/private/webserver.key.pem
+            touch /etc/ssl/docker-letsencrypt-configured
+        fi
     fi
 fi
